Enhancement: Extract concrete class
