+++ psalm-baseline.xml
@@ -1,11 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <files psalm-version="3.18.2@19aa905f7c3c7350569999a93c40ae91ae4e1626">
-  <file src="test/Unit/FactoryTest.php">
-    <MixedPropertyTypeCoercion occurrences="2">
-      <code>$rules</code>
-      <code>$rules</code>
-    </MixedPropertyTypeCoercion>
-  </file>
   <file src="test/Unit/RuleSet/AbstractRuleSetTestCase.php">
     <InternalClass occurrences="2">
       <code>FixerFactory::create()</code>
